config/gemini_client.py
import os
from dotenv import load_dotenv

# Load variables from .env file
load_dotenv()

# Placeholder for your actual Gemini client connection logic
def ask_gemini(prompt: str) -> str:
    """Uses the GEMINI_API_KEY from .env to make a mock call."""
    api_key = os.getenv("GEMINI_API_KEY")
    if not api_key:
        return "ERROR: GEMINI_API_KEY not set in .env file."
    
    # In a real app, you would use:
    # client = google.genai.Client(api_key=api_key)
    # response = client.models.generate_content(...)
    
    return "This is a mock response using the loaded API key."