requirements.txt
# Google Agent Development Kit
google-adk==0.5.2

# Gemini API and related dependencies
google-generativeai==0.4.1
google-ai-generativelanguage==0.6.12

# Environment variable management
python-dotenv==1.0.0

# Data validation
pydantic==2.12.0
pydantic-core==2.16.1

# Google Cloud dependencies (optional, for Vertex AI support)
google-cloud-aiplatform==1.43.0

# HTTP and async support
httpx==0.25.2
anyio==4.2.0

# API Server dependencies
fastapi==0.109.0
uvicorn==0.27.0
starlette==0.35.0

# JSON and serialization
typing-extensions==4.9.0

# Async support
aiofiles==23.2.1

# CLI support
click==8.1.7
typer==0.9.1

# Logging and monitoring
python-json-logger==2.0.7

# Development dependencies (optional)
pytest==7.4.4
pytest-asyncio==0.23.2
black==24.1.1
flake8==7.0.0
isort==5.13.2

